Move changesummary, hosts and replacebetweenmarkers to pkg
