+++ pkg/dependencyutils/dependencyinterfaces/Dependency.go
@@ -1,7 +1,7 @@
 package dependencyinterfaces
 
 import (
-	"github.com/asciich/asciichgolangpublic/changesummary"
+	"github.com/asciich/asciichgolangpublic/pkg/changesummary"
 	"github.com/asciich/asciichgolangpublic/pkg/files"
 	"github.com/asciich/asciichgolangpublic/pkg/parameteroptions"
 	"github.com/asciich/asciichgolangpublic/pkg/parameteroptions/authenticationoptions"
