+++ pkg/hosts/CommandExecutorHost_test.go
