+++ pkg/hosts/Host_test.go
@@ -141,7 +141,6 @@ func TestHost_GetDirectoryByPath(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		tt := tt
 		t.Run(
 			testutils.MustFormatAsTestname(tt),
 			func(t *testing.T) {
